64位机器上long unsigned int为64位，改为unsigned int类型

+++ include/md5.h
@@ -22,7 +22,8 @@
 extern "C" {
 #endif
 
-typedef long unsigned int uint32;
+//typedef long unsigned int uint32;
+typedef unsigned int uint32;
 typedef struct MD5Context MD5_CTX;
 
 #define md5byte unsigned char
